fix: prefer packaged skill evals
