+++ scripts/skill-eval/check_evals.mjs
@@ -2,10 +2,9 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 import { existsSync } from 'node:fs';
-import { mkdir, readdir, readFile, writeFile } from 'node:fs/promises';
+import { mkdir, readFile, writeFile } from 'node:fs/promises';
 import { basename, join, resolve } from 'node:path';
-
-const SKILL_PACKAGE_PREFIX = '@lynx-js/skill-';
+import { discoverSkillSuites } from './discover_suites.mjs';
 
 main().catch((error) => {
   console.error(error instanceof Error ? error.message : String(error));
@@ -81,79 +80,6 @@ function parseArgs(argv) {
     throw new Error(`Unknown argument: ${arg}`);
   }
   return args;
-}
-
-async function discoverSkillSuites(repoRoot) {
-  const skillsRoot = join(repoRoot, 'packages/skills');
-  if (!existsSync(skillsRoot)) {
-    throw new Error(`skills directory not found: ${skillsRoot}`);
-  }
-
-  const entries = await readdir(skillsRoot, { withFileTypes: true });
-  const sourceSuites = entries
-    .filter((entry) => entry.isDirectory())
-    .map((entry) => {
-      const skillPath = join(skillsRoot, entry.name);
-      return {
-        packageName: undefined,
-        evalPath: join(skillPath, 'evals'),
-        skillPath,
-      };
-    })
-    .filter((suite) => existsSync(join(suite.skillPath, 'SKILL.md')))
-    .sort((left, right) => left.skillPath.localeCompare(right.skillPath));
-
-  const sourcePackages = await getSourcePackageMap(sourceSuites);
-  const suitesBySkillPath = new Map(
-    sourceSuites.map((suite) => [suite.skillPath, suite]),
-  );
-
-  for (const dependencyName of await getSkillDependencies(repoRoot)) {
-    const sourceSuite = sourcePackages.get(dependencyName);
-    if (sourceSuite) {
-      sourceSuite.packageName = dependencyName;
-      continue;
-    }
-
-    const skillSlug = dependencyName.slice(SKILL_PACKAGE_PREFIX.length);
-    const skillPath = join(
-      repoRoot,
-      'node_modules',
-      ...dependencyName.split('/'),
-    );
-    suitesBySkillPath.set(skillPath, {
-      packageName: dependencyName,
-      evalPath: join(repoRoot, 'evals', skillSlug),
-      skillPath,
-    });
-  }
-
-  return Array.from(suitesBySkillPath.values()).sort((left, right) =>
-    String(left.packageName ?? left.skillPath).localeCompare(
-      String(right.packageName ?? right.skillPath),
-    ),
-  );
-}
-
-async function getSourcePackageMap(sourceSuites) {
-  const packages = new Map();
-  for (const suite of sourceSuites) {
-    const packageJsonPath = join(suite.skillPath, 'package.json');
-    if (!existsSync(packageJsonPath)) continue;
-    const packageJson = await readJson(packageJsonPath);
-    if (typeof packageJson.name === 'string') {
-      packages.set(packageJson.name, suite);
-    }
-  }
-  return packages;
-}
-
-async function getSkillDependencies(repoRoot) {
-  const packageJson = await readJson(join(repoRoot, 'package.json'));
-  const dependencies = packageJson.dependencies ?? {};
-  return Object.keys(dependencies)
-    .filter((name) => name.startsWith(SKILL_PACKAGE_PREFIX))
-    .sort();
 }
 
 function buildSingleSuite(args) {
