+++ scripts/skill-eval/run_task_online_all.mjs
@@ -2,11 +2,9 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 import { spawn } from 'node:child_process';
-import { existsSync } from 'node:fs';
-import { readdir, readFile } from 'node:fs/promises';
-import { basename, join, resolve } from 'node:path';
-
-const SKILL_PACKAGE_PREFIX = '@lynx-js/skill-';
+import { readFile } from 'node:fs/promises';
+import { join, resolve } from 'node:path';
+import { discoverSkillSuites, suiteSlug } from './discover_suites.mjs';
 
 main().catch((error) => {
   console.error(error instanceof Error ? error.message : String(error));
@@ -106,79 +104,6 @@ function parseArgs(argv) {
   return args;
 }
 
-async function discoverSkillSuites(repoRoot) {
-  const skillsRoot = join(repoRoot, 'packages/skills');
-  if (!existsSync(skillsRoot)) {
-    throw new Error(`skills directory not found: ${skillsRoot}`);
-  }
-
-  const entries = await readdir(skillsRoot, { withFileTypes: true });
-  const sourceSuites = entries
-    .filter((entry) => entry.isDirectory())
-    .map((entry) => {
-      const skillPath = join(skillsRoot, entry.name);
-      return {
-        evalPath: join(skillPath, 'evals'),
-        packageName: undefined,
-        skillPath,
-      };
-    })
-    .filter((suite) => existsSync(join(suite.skillPath, 'SKILL.md')))
-    .sort((left, right) => left.skillPath.localeCompare(right.skillPath));
-
-  const sourcePackages = await getSourcePackageMap(sourceSuites);
-  const suitesBySkillPath = new Map(
-    sourceSuites.map((suite) => [suite.skillPath, suite]),
-  );
-
-  for (const dependencyName of await getSkillDependencies(repoRoot)) {
-    const sourceSuite = sourcePackages.get(dependencyName);
-    if (sourceSuite) {
-      sourceSuite.packageName = dependencyName;
-      continue;
-    }
-
-    const skillSlug = dependencyName.slice(SKILL_PACKAGE_PREFIX.length);
-    const skillPath = join(
-      repoRoot,
-      'node_modules',
-      ...dependencyName.split('/'),
-    );
-    suitesBySkillPath.set(skillPath, {
-      evalPath: join(repoRoot, 'evals', skillSlug),
-      packageName: dependencyName,
-      skillPath,
-    });
-  }
-
-  return Array.from(suitesBySkillPath.values()).sort((left, right) =>
-    String(left.packageName ?? left.skillPath).localeCompare(
-      String(right.packageName ?? right.skillPath),
-    ),
-  );
-}
-
-async function getSourcePackageMap(sourceSuites) {
-  const packages = new Map();
-  for (const suite of sourceSuites) {
-    const packageJsonPath = join(suite.skillPath, 'package.json');
-    if (!existsSync(packageJsonPath)) continue;
-    const packageJson = await readJson(packageJsonPath);
-    if (typeof packageJson.name === 'string') {
-      packages.set(packageJson.name, suite);
-    }
-  }
-  return packages;
-}
-
-async function getSkillDependencies(repoRoot) {
-  const packageJson = await readJson(join(repoRoot, 'package.json'));
-  const dependencies = packageJson.dependencies ?? {};
-  return Object.keys(dependencies)
-    .filter((name) => name.startsWith(SKILL_PACKAGE_PREFIX))
-    .sort();
-}
-
 async function resolveEvalIds(suite, args) {
   if (args.evalIds) {
     return args.evalIds
@@ -197,13 +122,6 @@ async function resolveEvalIds(suite, args) {
     .map((evalItem) => evalItem.id)
     .filter((id) => Number.isInteger(id))
     .slice(0, args.maxEvalsPerSkill);
-}
-
-function suiteSlug(suite) {
-  if (suite.packageName?.startsWith(SKILL_PACKAGE_PREFIX)) {
-    return suite.packageName.slice(SKILL_PACKAGE_PREFIX.length);
-  }
-  return basename(suite.skillPath);
 }
 
 function parseList(value) {
